requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
confusable-homoglyphs==3.3.1
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-registration==3.4
django-rq==2.10.2
djangorestframework==3.15.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.7
rq==1.15.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
